fix(ram): the `ram_resource_permission` dataSource quality improvement
