+++ docs/data-sources/ram_resource_permission.md
@@ -2,23 +2,21 @@
 subcategory: "Resource Access Manager (RAM)"
 layout: "huaweicloud"
 page_title: "HuaweiCloud: huaweicloud_ram_resource_permission"
-description: |
-  Use this data source to get the permission for the specified version.
+description: |-
+  Use this data source to get the RAM resource permission details within HuaweiCloud.
 ---
 
 # huaweicloud_ram_resource_permission
 
-Use this data source to get the permission for the specified version.
+Use this data source to get the RAM resource permission details within HuaweiCloud.
 
 ## Example Usage
 
 ```hcl
 variable "permission_id" {}
-variable "permission_version" {}
 
 data "huaweicloud_ram_resource_permission" "test" {
-  resource_type      = var.permission_id
-  permission_version = var.permission_version
+  permission_id = var.permission_id
 }
 ```
 
@@ -26,7 +24,7 @@ data "huaweicloud_ram_resource_permission" "test" {
 
 The following arguments are supported:
 
-* `permission_id` - (Required, String) Specifies the id of RAM permission.
+* `permission_id` - (Required, String) Specifies the ID of RAM permission.
 
 * `permission_version` - (Optional, String) Specifies the version of RAM permission.
   Not setting this value will return the default version permission.
@@ -37,23 +35,24 @@ In addition to all arguments above, the following attributes are exported:
 
 * `id` - The data source ID.
 
-* `permission` - Indicates the list of the RAM permissions
-  The [permission](#RAM_Permission) structure is documented below.
+* `permission` - Indicates the permission details.
 
-<a name="RAM_Permission"></a>
+  The [permission](#permission_struct) structure is documented below.
+
+<a name="permission_struct"></a>
 The `permission` block supports:
 
-* `id` - Indicates the id of RAM permission.
+* `id` - Indicates the ID of RAM permission.
 
 * `name` - Indicates the name of RAM permission.
 
 * `resource_type` - Indicates the resource type of RAM permission.
 
-* `is_resource_type_default` - Whether the RAM permission resource type is default.
-
 * `content` - Impact and actions allowed by the permission.
 
-* `created_at` - Indicates the RAM permission create time.
+* `is_resource_type_default` - Whether the RAM permission resource type is default.
+
+* `created_at` - Indicates the RAM permission creation time.
 
 * `updated_at` - Indicates the RAM permission last update time.
